Extract cost checker from inliner.
